Solved another tEnvelope bug. [Note: port update code to drumbox.] [Note2: the attack/decayPhase and attack/decayInc class variables are now all floats in tEnvelope. should determine whether this is okay.]

+++ CAudio/Source/CAudioLibrary.h
@@ -53,7 +53,7 @@ typedef struct _tEnvelope {
     
     float gain;
     
-    uint32_t attackPhase, decayPhase;
+    float attackPhase, decayPhase;
     
     float(*tick)(struct _tEnvelope *self);
     int(*on)(struct _tEnvelope *self, float velocity);
